move SpotifyPlayer
